Remove unused test configuration
