+++ apps/meteor/package.json
@@ -16,9 +16,9 @@
 		"start": "meteor",
 		"build:ci": "METEOR_DISABLE_OPTIMISTIC_CACHING=1 meteor build --server-only --directory /tmp/dist",
 		"dev": "meteor --exclude-archs \"web.browser.legacy, web.cordova\"",
-		"dsv": "meteor npm run dev",
-		"ha": "meteor npm run ha:start",
-		"ms": "TRANSPORTER=${TRANSPORTER:-TCP} meteor npm run dev",
+		"dsv": "meteor yarn dev",
+		"ha": "meteor yarn ha:start",
+		"ms": "TRANSPORTER=${TRANSPORTER:-TCP} meteor yarn dev",
 		"ha:start": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' ts-node .scripts/run-ha.ts main",
 		"ha:add": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' ts-node .scripts/run-ha.ts instance",
 		"migration:add": "ts-node-transpile-only --skip-project .scripts/make-migration.ts",
@@ -31,25 +31,25 @@
 		"stylelint": "stylelint \"app/**/*.css\" \"client/**/*.css\" \"app/**/*.less\" \"client/**/*.less\" \"ee/**/*.less\"",
 		"stylelint:fix": "stylelint --fix \"app/**/*.css\" \"client/**/*.css\" \"app/**/*.less\" \"client/**/*.less\" \"ee/**/*.less\"",
 		"typecheck": "cross-env NODE_OPTIONS=\"--max-old-space-size=8192\" tsc -p tsconfig.typecheck.json",
-		"deploy": "npm run build && pm2 startOrRestart pm2.json",
+		"deploy": "yarn build && pm2 startOrRestart pm2.json",
 		"coverage": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' nyc -r html mocha --config ./.mocharc.js",
 		"test:e2e": "playwright test",
 		"test:e2e:federation": "playwright test --config=playwright-federation.config.ts",
 		"test:e2e:nyc": "nyc report --reporter=lcov",
 		"testapi": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha --config ./.mocharc.api.js",
-		"testunit": "npm run .testunit:definition && npm run .testunit:client && npm run .testunit:server:cov",
+		"testunit": "yarn .testunit:definition && yarn .testunit:client && yarn .testunit:server:cov",
 		".testunit:server": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha --config ./.mocharc.js",
 		".testunit:server:cov": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' nyc -r text -r lcov mocha --config ./.mocharc.js",
-		".testunit:client": "jest --silent && TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha --config ./.mocharc.client.js --exit",
+		".testunit:client": "jest",
 		".testunit:definition": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha --config ./.mocharc.definition.js",
 		"testunit-watch": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' mocha --watch --config ./.mocharc.js",
-		"test": "npm run testapi && npm run testui",
+		"test": "yarn testunit && yarn testapi",
 		"translation-diff": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' ts-node .scripts/translation-diff.ts",
 		"translation-check": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' ts-node .scripts/translation-check.ts",
 		"translation-fix-order": "TS_NODE_COMPILER_OPTIONS='{\"module\": \"commonjs\"}' ts-node .scripts/translation-fix-order.ts",
 		"version": "node .scripts/version.js",
 		"set-version": "node .scripts/set-version.js",
-		"release": "meteor npm run set-version --silent",
+		"release": "meteor yarn set-version --silent",
 		"storybook": "cross-env NODE_OPTIONS=--max-old-space-size=8192 start-storybook -p 6006 --no-version-updates",
 		"prepare": "node playwright.prepare.mjs",
 		"docker:start": "docker-compose up"
