--- apps/meteor/tests/setup/registerWebApiMocks.ts
@@ -1,5 +1,0 @@
-import { enableBlobUrlsMock } from '../mocks/client/blobUrls';
-import { enableJsdom } from '../mocks/client/jsdom';
-
-enableJsdom();
-enableBlobUrlsMock();
